requirements.txt
alembic==1.8.1
anyio==3.6.2
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
# charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchgod==0.8.2
websockets==10.4
